build basic project
